pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bio-ml-agent"
version = "0.2.0"
description = "Bio-ML Agent - Yerel LLM destekli Biyomühendislik ML Proje Asistanı"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name="Bio-ML Agent Contributors" }
]
dependencies = [
    # Çekirdek (Core)
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "duckduckgo-search>=7.0.0",
    
    # Yerel LLM
    "ollama>=0.4.0",
    
    # RAG (Vector DB)
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",

    # Pydantic (Ayar Denetimi - Task 5)
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",

    # ML / Veri Bilimi Çekirdeği
    "scikit-learn>=1.2.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
# UI ve Web Sunucuları
ui = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gradio>=4.40.0,<5.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
]

# Bulut LLM Sağlayıcıları (Remote Backend)
cloud = [
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-genai>=0.5.0",
]

# Görselleştirme
viz = [
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
]

# WhatsApp Entegrasyonu
whatsapp = [
    "twilio>=8.0.0",
]

# Derin Öğrenme (Deep Learning)
dl = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.65.0",
    "Pillow>=9.0.0",
]

# Açıklanabilir Yapay Zeka (XAI)
xai = [
    "shap>=0.42.0",
    "lime>=0.2.0.1",
]

# Test ve Geliştirme
test = [
    "pytest>=7.0.0",
]

# Hepsi
all = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gradio>=4.40.0,<5.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
    "openai>=1.0.0",
    "anthropic>=0.30.0",
    "google-genai>=0.5.0",
    "matplotlib>=3.6.0",
    "seaborn>=0.12.0",
    "plotly>=5.15.0",
    "twilio>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "tqdm>=4.65.0",
    "Pillow>=9.0.0",
    "shap>=0.42.0",
    "lime>=0.2.0.1",
    "pytest>=7.0.0",
]

[project.scripts]
bio-ml-agent = "agent:main"
